Transform class is created, that transforms the data using vector method, across the channel/Time.
However, the transform and masking are removed from the nowcast_dataset class. Because, masking is needed  after transform and inverse transform. On the other hand, transform is needed before providing to the model, while inverse transform is needed before saving to a zarr. To keep consistancy, do transformation+masking syncronously.